LayoutTests/fast/shadow-dom/Node-prototype-cloneNode.html

<!DOCTYPE html>
<html>
<head>
<title>Shadow DOM: Extensions to Node interface</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="Invoking the cloneNode() method on a ShadowRoot instance must always throw a NotSupportedError.">
<link rel="help" href="http://w3c.github.io/webcomponents/spec/shadow/#the-shadowroot-interface">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='../../resources/testharness.css'>
</head>
<body>
<div id="log"></div>
<script>

function testCloneNode(mode) {
    test(function () {
        assert_throws({'name': 'NotSupportedError'}, function () {
            var element = document.createElement('div');
            var shadowRoot = element.attachShadow({mode: mode});
            shadowRoot.cloneNode(false);
        }, 'cloneNode(false) on a shadow root in ' + mode + ' mode must throw a NotSupportedError');

        assert_throws({'name': 'NotSupportedError'}, function () {
            var element = document.createElement('div');
            var shadowRoot = element.attachShadow({mode: mode});
            shadowRoot.cloneNode(false);
        }, 'cloneNode(true) on a closed shadow root must throw a NotSupportedError');

    }, 'cloneNode on a shadow root in ' + mode + ' mode must throw a NotSupportedError');
}

testCloneNode('open');
testCloneNode('closed');

test(function () {
    var element = document.createElement('div');
    var shadowRoot = element.attachShadow({mode: 'open'});

    assert_equals(element.cloneNode(false).shadowRoot, null, 'cloneNode(false) on an open shadow host mode must clone its shadow root');
    assert_equals(element.cloneNode(true).shadowRoot, null, 'cloneNode(true) on an open shadow host mode must clone its shadow root');
}, 'cloneNode on an open shadow root must throw a NotSupportedError');

</script>
</body>
</html>
